Separate out Ensemble Copula Coupling plugins into a separate file, as this will require extending beyond the initial scope of the application to ensemble calibration output.
